add real time app and move testing
